+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## DEV
 
+ * Improve the text for repeated events (see #175).
  * Allow to overwrite the page title and description in news and events (see #161).
  * Show all root languages by default in the meta editor (see contao/core#6254).
  * Try to preserve existing .htaccess entries when installing the web directory (see #160).